[WIP][Discussion] App initiates a getAllProducts from first app load
