+++ client/src/components/SearchBar.jsx
@@ -2,26 +2,12 @@ import { LuSlidersHorizontal } from "react-icons/lu";
 import { FiSearch } from "react-icons/fi";
 import { CgCheck } from "react-icons/cg";
 import { Link } from "react-router-dom";
-import { useEffect, useState } from "react";
-import { allProducts } from "../app/products/products";
-import { useDispatch } from "react-redux";
+import { useState } from "react";
 
 const SearchBar = () => {
-
-  const dispatch = useDispatch();
   const [input, setInput] = useState("");
 
-  useEffect(() => {
-    const getAllProducts = async () => {
-      const response = await fetch("http://localhost:3000/nike/getAll");
-      const data = await response.json();
-      dispatch(allProducts(data));
-    };
-    getAllProducts();
-  }, [location.pathname]);
-
-
-  //   return value && product && product.itemName && product.itemName.toLowerCase().includes(value);
+  //  return value && product && product.itemName && product.itemName.toLowerCase().includes(value);
 
 
   const handleChange = (value) => {
